requirements.txt
PySide6>=6.5.0
pymavlink>=2.4.37
pyserial>=3.5
numpy>=1.24.0
pydantic>=2.0.0
pytest>=7.3.1
pytest-qt>=4.2.0
pytest-asyncio>=0.21.0
black>=23.3.0  # Code-Formatierung
mypy>=1.3.0   # Statische Typ-Überprüfung
flake8>=6.0.0 # Linting 